a bunch of explicit (float) casts etc to appease MSVC

+++ Examples/IPlugEffect/IPlugEffect_controls.h
@@ -226,11 +226,11 @@ public:
       }
       else if (mShape == 1)
       {
-        float pad1 = (mRECT.W() / 2.0) * (1.0 - mValue);
-        float pad2 = (mRECT.H() / 2.0) * (1.0 - mValue);
+        float pad1 = (mRECT.W() / 2.f) * (1.f - (float) mValue);
+        float pad2 = (mRECT.H() / 2.f) * (1.f - (float) mValue);
         IRECT size1 = mRECT.GetPadded(pad1, pad2, -pad1, -pad2);
-        pad1 = (size1.W() / 2.0) * (1.0 - mValue);
-        pad2 = (size1.H() / 2.0) * (1.0 - mValue);
+        pad1 = (size1.W() / 2.f) * (1.f - (float) mValue);
+        pad2 = (size1.H() / 2.f) * (1.f - (float) mValue);
         IRECT size2 = size1.GetPadded(pad1, pad2, -pad1, -pad2);
         g.PathRect(size1);
         g.PathRect(size2);
